% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsra.R
\name{rsra}
\alias{rsra}
\title{Query and download SRA files from NCBI}
\usage{
rsra(idList, outdir = getwd(), progress = TRUE, location = c("AWS", "NCBI"))
}
\arguments{
\item{idList}{a list of cases to download.}

\item{outdir}{output directory, default is working directory.}

\item{progress}{if \code{TRUE}, show 'wget' download progress.}

\item{location}{one of "AWS" or "NCBI" for download server.
"GCAP" is not available due to its limit.}
}
\value{
Nothing
}
\description{
Query and download SRA files from NCBI
}
