% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsra.R
\name{rsra}
\alias{rsra}
\title{Query and download SRA files from NCBI}
\usage{
rsra(
  idList,
  prefetch = NULL,
  outdir = getwd(),
  progress = TRUE,
  location = c("AWS", "NCBI"),
  opts = "-r yes -C yes"
)
}
\arguments{
\item{idList}{a list of cases to download.}

\item{prefetch}{if not \code{NULL}, should be the path to the prefetch program.}

\item{outdir}{output directory, default is working directory.}

\item{progress}{if \code{TRUE}, show 'wget' download progress.}

\item{location}{one of "AWS" or "NCBI" for download server.
"GCAP" is not available due to its limit.}

\item{opts}{options work with \code{prefetch} (expects \code{-p}) when \code{prefetch} is not \code{NULL}.}
}
\value{
Nothing
}
\description{
Query and download SRA files from NCBI
}
